requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
chromadb
python-dotenv
pypdf
langgraph
fastapi
uvicorn
streamlit
mermaid-python
pygraphviz
pydantic
python-multipart
tree

